+++ plugin/web/shop_state_request_handler.rb
@@ -1,0 +1,12 @@
+module AresMUSH
+  module OsrRpg
+    class ShopStateRequestHandler
+      def handle(request)
+        error = Website.check_login(request)
+        return { error: error } if error
+
+        ShopHelper.state_for_web(request.enactor)
+      end
+    end
+  end
+end